Move MCP deployment files from app/ to odh/ directory
MCP deployment pages, hooks, API services, types, and utils are
ODH-specific code that belongs in upstream/frontend/src/odh/, not
in the upstream-synced app/ directory. This move keeps them as part
of the existing modelRegistry federated module with no changes to
containers, build processes, or dev servers.
