+++ packages/model-registry/upstream/frontend/src/odh/McpDeploymentsWrapper.tsx
@@ -11,7 +11,7 @@ import { ThemeProvider, Theme } from 'mod-arch-kubeflow';
 import { Bullseye } from '@patternfly/react-core';
 import { BFF_API_VERSION, URL_PREFIX } from '~/app/utilities/const';
 import { AppContext } from '~/app/context/AppContext';
-import McpDeploymentsRoutes from '~/app/pages/mcpDeployments/McpDeploymentsRoutes';
+import McpDeploymentsRoutes from '~/odh/pages/mcpDeployments/McpDeploymentsRoutes';
 import NotificationListener from '~/odh/components/NotificationListener';
 import OdhDevFeatureFlagOverridesProvider from '~/odh/components/OdhDevFeatureFlagOverridesProvider';
