+++ packages/model-registry/upstream/frontend/src/odh/api/mcpCatalogDeployment/service.ts
@@ -13,7 +13,7 @@ import {
   McpDeployment,
   McpDeploymentCreateRequest,
   McpDeploymentUpdateRequest,
-} from '~/app/mcpDeploymentTypes';
+} from '~/odh/types/mcpDeploymentTypes';
 
 export type McpServerAvailabilityResponse = {
   available: boolean;
