+++ packages/model-registry/upstream/frontend/src/odh/components/McpDeployModal.tsx
@@ -21,14 +21,14 @@ import {
   isValidK8sName,
 } from '@odh-dashboard/internal/concepts/k8s/utils';
 import NamespaceSelectorFieldWrapper from '~/odh/components/NamespaceSelectorFieldWrapper';
-import useMcpServerConverter from '~/app/hooks/mcpCatalogDeployment/useMcpServerConverter';
+import useMcpServerConverter from '~/odh/hooks/useMcpServerConverter';
 import K8sNameDescriptionField from '~/concepts/k8s/K8sNameDescriptionField/K8sNameDescriptionField';
 import { K8sNameDescriptionFieldData } from '~/concepts/k8s/K8sNameDescriptionField/types';
 import { MAX_K8S_NAME_LENGTH } from '~/concepts/k8s/K8sNameDescriptionField/utils';
-import { createMcpDeployment, updateMcpDeployment } from '~/app/api/mcpCatalogDeployment/service';
+import { createMcpDeployment, updateMcpDeployment } from '~/odh/api/mcpCatalogDeployment/service';
 import { mcpDeploymentsUrl } from '~/app/routes/mcpCatalog/mcpCatalog';
-import { mcpServerCRToYaml } from '~/app/utils/mcpServerYaml';
-import { McpDeployment } from '~/app/mcpDeploymentTypes';
+import { mcpServerCRToYaml } from '~/odh/utils/mcpServerYaml';
+import { McpDeployment } from '~/odh/types/mcpDeploymentTypes';
 
 type McpDeployModalProps = {
   isOpen?: boolean;
