+++ packages/model-registry/upstream/frontend/src/odh/pages/mcpDeployments/McpDeploymentsEmptyState.tsx
