+++ packages/model-registry/upstream/frontend/src/odh/pages/mcpDeployments/__tests__/DeleteMcpDeploymentModal.spec.tsx
@@ -2,11 +2,11 @@ import '@testing-library/jest-dom';
 import * as React from 'react';
 import { render, screen, fireEvent, waitFor } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
-import { deleteMcpDeployment } from '~/app/api/mcpDeploymentService';
-import DeleteMcpDeploymentModal from '~/app/pages/mcpDeployments/DeleteMcpDeploymentModal';
+import { deleteMcpDeployment } from '~/odh/api/mcpDeploymentService';
+import DeleteMcpDeploymentModal from '~/odh/pages/mcpDeployments/DeleteMcpDeploymentModal';
 import { createMockDeployment } from './mcpDeploymentTestUtils';
 
-jest.mock('~/app/api/mcpDeploymentService', () => ({
+jest.mock('~/odh/api/mcpDeploymentService', () => ({
   deleteMcpDeployment: jest.fn(),
 }));
